Move action to VersionResource
