+++ app/avo/resources/rubygem_resource.rb
@@ -16,7 +16,6 @@ class RubygemResource < Avo::BaseResource
 
   action AddOwner
   action ReleaseReservedNamespace
-  action RestoreRubygemVersion
   action YankRubygem
   action ReserveNamespace
   action UploadInfoFile
